fix: error with number value in calculateLength (#300)
It is possible to have a number in an input rather than a string if you have the input set to type="number" in this case we were getting a "value is not iterable" error when the value was a number. Added an additional check to make sure the value is a string.